socketio-backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==4.3.1
python-engineio==3.9.0
python-multipart==0.0.6
